Chuck Norris -> Vin Diesel
